requirements.txt
Django==4.2.7
Pillow==10.0.1
django-crispy-forms==2.0
crispy-tailwind==0.5.0
python-decouple==3.8
django-jazzmin==2.6.0
django-environ==0.11.2
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
whitenoise==6.6.0